Introduce WithImageConstrained
