+++ ts/editor/ImageHandleFloat.svelte
@@ -11,7 +11,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
     import { floatNoneIcon, floatLeftIcon, floatRightIcon } from "./icons";
 
-    export let activeImage: HTMLImageElement;
+    export let image: HTMLImageElement;
     export let isRtl: boolean;
 
     const leftValues = {
@@ -33,20 +33,18 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
     <ButtonGroupItem>
         <IconButton
             tooltip={tr.editingFloatNone()}
-            active={activeImage.style.float === "" ||
-                activeImage.style.float === "none"}
+            active={image.style.float === "" || image.style.float === "none"}
             flipX={isRtl}
-            on:click={() => (activeImage.style.float = "")}
-            >{@html floatNoneIcon}</IconButton
+            on:click={() => (image.style.float = "")}>{@html floatNoneIcon}</IconButton
         >
     </ButtonGroupItem>
 
     <ButtonGroupItem>
         <IconButton
             tooltip={inlineStart.label}
-            active={activeImage.style.float === inlineStart.position}
+            active={image.style.float === inlineStart.position}
             flipX={isRtl}
-            on:click={() => (activeImage.style.float = inlineStart.position)}
+            on:click={() => (image.style.float = inlineStart.position)}
             >{@html floatLeftIcon}</IconButton
         >
     </ButtonGroupItem>
@@ -54,9 +52,9 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
     <ButtonGroupItem>
         <IconButton
             tooltip={inlineEnd.label}
-            active={activeImage.style.float === inlineEnd.position}
+            active={image.style.float === inlineEnd.position}
             flipX={isRtl}
-            on:click={() => (activeImage.style.float = inlineEnd.position)}
+            on:click={() => (image.style.float = inlineEnd.position)}
             >{@html floatRightIcon}</IconButton
         >
     </ButtonGroupItem>
